LayoutTests/fast/device-orientation/device-motion-request-permission-granted-expected.txt

CONSOLE MESSAGE: line 12: No device motion or orientation events will be fired until permission has been requested and granted.
Received device orientation & motion access request for security origin "".
Basic testing for DeviceMotionEvent.requestPermission().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is "granted"
PASS result is "granted"
PASS successfullyParsed is true

TEST COMPLETE
